feat: Upgrade ContosoUniversity from .NET Framework 4.8 to .NET 10
- Convert ContosoUniversity.csproj to SDK-style format targeting net10.0
- Replace ASP.NET MVC 5 (System.Web) with ASP.NET Core MVC
- Add Program.cs replacing Global.asax.cs application startup
- Add appsettings.json replacing Web.config configuration
- Migrate packages.config to PackageReference format
- Update EF Core from 3.1.32 to 9.0.5 (latest compatible with net10.0)
- Replace System.Messaging (MSMQ) with in-memory ConcurrentQueue
- Update all controllers to use DI constructor injection
- Update all controllers to use ASP.NET Core MVC patterns
- Update views to use tag helpers instead of HtmlHelpers where appropriate
- Remove App_Start folder (RouteConfig, BundleConfig, FilterConfig)
- Remove Web.config, Global.asax, packages.config

+++ ContosoUniversity/Controllers/InstructorsController.cs
@@ -1,19 +1,20 @@
-using System;
-using System.Collections.Generic;
-using Microsoft.EntityFrameworkCore;
-using System.Linq;
-using System.Net;
-using System.Web.Mvc;
 using ContosoUniversity.Data;
 using ContosoUniversity.Models;
 using ContosoUniversity.Models.SchoolViewModels;
+using ContosoUniversity.Services;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 
 namespace ContosoUniversity.Controllers
 {
     public class InstructorsController : BaseController
     {
-        // GET: Instructors - All roles can view
-        public ActionResult Index(int? id, int? courseID)
+        public InstructorsController(SchoolContext context, NotificationService notificationService)
+            : base(context, notificationService)
+        {
+        }
+
+        public IActionResult Index(int? id, int? courseID)
         {
             var viewModel = new InstructorIndexData();
             viewModel.Instructors = db.Instructors
@@ -33,30 +34,28 @@ namespace ContosoUniversity.Controllers
             if (courseID != null)
             {
                 ViewBag.CourseID = courseID.Value;
-                viewModel.Enrollments = viewModel.Courses.Where(
+                viewModel.Enrollments = viewModel.Courses!.Where(
                     x => x.CourseID == courseID).Single().Enrollments;
             }
 
             return View(viewModel);
         }
 
-        // GET: Instructors/Details/5 - All roles can view details
-        public ActionResult Details(int? id)
+        public IActionResult Details(int? id)
         {
             if (id == null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return BadRequest();
             }
-            Instructor instructor = db.Instructors.Find(id);
+            Instructor? instructor = db.Instructors.Find(id);
             if (instructor == null)
             {
-                return HttpNotFound();
+                return NotFound();
             }
             return View(instructor);
         }
 
-        // GET: Instructors/Create
-        public ActionResult Create()
+        public IActionResult Create()
         {
             var instructor = new Instructor();
             instructor.CourseAssignments = new List<CourseAssignment>();
@@ -64,10 +63,9 @@ namespace ContosoUniversity.Controllers
             return View(instructor);
         }
 
-        // POST: Instructors/Create
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "LastName,FirstMidName,HireDate,OfficeAssignment")] Instructor instructor, string[] selectedCourses)
+        public IActionResult Create([Bind("LastName,FirstMidName,HireDate,OfficeAssignment")] Instructor instructor, string[]? selectedCourses)
         {
             if (selectedCourses != null)
             {
@@ -82,34 +80,32 @@ namespace ContosoUniversity.Controllers
             {
                 db.Instructors.Add(instructor);
                 db.SaveChanges();
-                
-                // Send notification for instructor creation
+
                 SendEntityNotification("Instructor", instructor.ID.ToString(), EntityOperation.CREATE);
-                
+
                 return RedirectToAction("Index");
             }
             PopulateAssignedCourseData(instructor);
             return View(instructor);
         }
 
-        // GET: Instructors/Edit/5
-        public ActionResult Edit(int? id)
+        public IActionResult Edit(int? id)
         {
             if (id == null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return BadRequest();
             }
-            Instructor instructor = db.Instructors
+            Instructor? instructor = db.Instructors
                 .Include(i => i.OfficeAssignment)
                 .Include(i => i.CourseAssignments)
                     .ThenInclude(c => c.Course)
                 .Where(i => i.ID == id)
-                .Single();
-            PopulateAssignedCourseData(instructor);
+                .SingleOrDefault();
             if (instructor == null)
             {
-                return HttpNotFound();
+                return NotFound();
             }
+            PopulateAssignedCourseData(instructor);
             return View(instructor);
         }
 
@@ -130,14 +126,13 @@ namespace ContosoUniversity.Controllers
             ViewBag.Courses = viewModel;
         }
 
-        // POST: Instructors/Edit/5
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit(int? id, string[] selectedCourses)
+        public async Task<IActionResult> EditPost(int? id, string[]? selectedCourses)
         {
             if (id == null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return BadRequest();
             }
             var instructorToUpdate = db.Instructors
                .Include(i => i.OfficeAssignment)
@@ -146,12 +141,12 @@ namespace ContosoUniversity.Controllers
                .Where(i => i.ID == id)
                .Single();
 
-            if (TryUpdateModel(instructorToUpdate, "",
-               new string[] { "LastName", "FirstMidName", "HireDate", "OfficeAssignment" }))
+            if (await TryUpdateModelAsync<Instructor>(instructorToUpdate, "",
+               i => i.LastName, i => i.FirstMidName, i => i.HireDate, i => i.OfficeAssignment!))
             {
                 try
                 {
-                    if (String.IsNullOrWhiteSpace(instructorToUpdate.OfficeAssignment.Location))
+                    if (String.IsNullOrWhiteSpace(instructorToUpdate.OfficeAssignment?.Location))
                     {
                         instructorToUpdate.OfficeAssignment = null;
                     }
@@ -159,10 +154,48 @@ namespace ContosoUniversity.Controllers
                     UpdateInstructorCourses(selectedCourses, instructorToUpdate);
 
                     db.SaveChanges();
-                    
-                    // Send notification for instructor update
+
                     SendEntityNotification("Instructor", instructorToUpdate.ID.ToString(), EntityOperation.UPDATE);
 
+                    return RedirectToAction("Index");
+                }
+                catch (Exception)
+                {
+                    ModelState.AddModelError("", "Unable to save changes. Try again, and if the problem persists, see your system administrator.");
+                }
+            }
+            PopulateAssignedCourseData(instructorToUpdate);
+            return View("Edit", instructorToUpdate);
+        }
+
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public IActionResult Edit(int? id, string[]? selectedCourses)
+        {
+            if (id == null)
+            {
+                return BadRequest();
+            }
+            var instructorToUpdate = db.Instructors
+               .Include(i => i.OfficeAssignment)
+               .Include(i => i.CourseAssignments)
+                   .ThenInclude(c => c.Course)
+               .Where(i => i.ID == id)
+               .Single();
+
+            if (ModelState.IsValid)
+            {
+                try
+                {
+                    if (String.IsNullOrWhiteSpace(instructorToUpdate.OfficeAssignment?.Location))
+                    {
+                        instructorToUpdate.OfficeAssignment = null;
+                    }
+
+                    UpdateInstructorCourses(selectedCourses, instructorToUpdate);
+                    db.SaveChanges();
+
+                    SendEntityNotification("Instructor", instructorToUpdate.ID.ToString(), EntityOperation.UPDATE);
                     return RedirectToAction("Index");
                 }
                 catch (Exception)
@@ -174,7 +207,7 @@ namespace ContosoUniversity.Controllers
             return View(instructorToUpdate);
         }
 
-        private void UpdateInstructorCourses(string[] selectedCourses, Instructor instructorToUpdate)
+        private void UpdateInstructorCourses(string[]? selectedCourses, Instructor instructorToUpdate)
         {
             if (selectedCourses == null)
             {
@@ -183,8 +216,7 @@ namespace ContosoUniversity.Controllers
             }
 
             var selectedCoursesHS = new HashSet<string>(selectedCourses);
-            var instructorCourses = new HashSet<int>
-                (instructorToUpdate.CourseAssignments.Select(c => c.Course.CourseID));
+            var instructorCourses = new HashSet<int>(instructorToUpdate.CourseAssignments.Select(c => c.Course.CourseID));
             foreach (var course in db.Courses)
             {
                 if (selectedCoursesHS.Contains(course.CourseID.ToString()))
@@ -196,66 +228,57 @@ namespace ContosoUniversity.Controllers
                 }
                 else
                 {
-
                     if (instructorCourses.Contains(course.CourseID))
                     {
-                        CourseAssignment courseToRemove = instructorToUpdate.CourseAssignments.SingleOrDefault(i => i.CourseID == course.CourseID);
-                        db.Entry(courseToRemove).State = EntityState.Deleted;
+                        CourseAssignment? courseToRemove = instructorToUpdate.CourseAssignments.SingleOrDefault(i => i.CourseID == course.CourseID);
+                        if (courseToRemove != null)
+                            db.Entry(courseToRemove).State = EntityState.Deleted;
                     }
                 }
             }
         }
 
-        // GET: Instructors/Delete/5
-        public ActionResult Delete(int? id)
+        public IActionResult Delete(int? id)
         {
             if (id == null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return BadRequest();
             }
-            Instructor instructor = db.Instructors.Find(id);
+            Instructor? instructor = db.Instructors.Find(id);
             if (instructor == null)
             {
-                return HttpNotFound();
+                return NotFound();
             }
             return View(instructor);
         }
 
-        // POST: Instructors/Delete/5 - Only admins can delete instructors
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
+        public IActionResult DeleteConfirmed(int id)
         {
-            Instructor instructor = db.Instructors
+            Instructor? instructor = db.Instructors
               .Include(i => i.OfficeAssignment)
               .Where(i => i.ID == id)
-              .Single();
+              .SingleOrDefault();
 
-            db.Instructors.Remove(instructor);
-
-            var department = db.Departments
-                .Where(d => d.InstructorID == id)
-                .SingleOrDefault();
-            if (department != null)
+            if (instructor != null)
             {
-                department.InstructorID = null;
+                db.Instructors.Remove(instructor);
+
+                var department = db.Departments
+                    .Where(d => d.InstructorID == id)
+                    .SingleOrDefault();
+                if (department != null)
+                {
+                    department.InstructorID = null;
+                }
+
+                db.SaveChanges();
+
+                SendEntityNotification("Instructor", id.ToString(), EntityOperation.DELETE);
             }
 
-            db.SaveChanges();
-            
-            // Send notification for instructor deletion
-            SendEntityNotification("Instructor", id.ToString(), EntityOperation.DELETE);
-            
             return RedirectToAction("Index");
-        }
-
-        protected override void Dispose(bool disposing)
-        {
-            if (disposing)
-            {
-                db.Dispose();
-            }
-            base.Dispose(disposing);
         }
     }
 }
